Add comments and fix a few bugs
Comment the functions. Don't disable the update timer in the UI when the
internet turns off, since then we won't update the UI to show that.

Fix handling of state in Python and javascript (make it consistently 0 for
off and 1 for on).

Fix incorrect initial remaining time (the pending time from the 'state'
table was being subtracted twice).

Signed-off-by: Simon Glass <sjg@chromium.org>